Convert the rest of the tools
