+++ lib/tooling/base-tool.interface.ts
@@ -24,10 +24,11 @@
 
 import {Library} from '../../types/libraries/libraries.interfaces';
 import {Tool, ToolResult} from '../../types/tool.interfaces';
+import {PropertyGetter} from '../properties.interfaces';
 
 export type ToolEnv = {
-    ceProps: (key: string, defaultValue?: any) => string | boolean | number | undefined;
-    compilerProps: (key: string, defaultValue?: any) => string | boolean | number | undefined;
+    ceProps: PropertyGetter;
+    compilerProps: PropertyGetter;
 };
 
 export interface ITool extends Tool {
