+++ lib/tooling/testing-tool.ts
@@ -1,4 +1,4 @@
-// Copyright (c) 2018, Compiler Explorer Authors
+// Copyright (c) 2023, Compiler Explorer Authors
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -22,23 +22,10 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-import fs from 'fs-extra';
-
 import {BaseTool} from './base-tool';
 
-export class PaholeTool extends BaseTool {
+export class TestingTool extends BaseTool {
     static get key() {
-        return 'pahole-tool';
-    }
-
-    async runTool(compilationInfo, inputFilepath, args) {
-        if (!compilationInfo.filters.binary && !compilationInfo.filters.binaryObject) {
-            return this.createErrorResponse('Pahole requires a binary output');
-        }
-
-        if (await fs.pathExists(compilationInfo.executableFilename)) {
-            return super.runTool(compilationInfo, compilationInfo.executableFilename, args);
-        }
-        return super.runTool(compilationInfo, compilationInfo.outputFilename, args);
+        return 'testing-tool';
     }
 }
